src/phantom/automated_attack_system.py
import nmap
import logging
from typing import Dict, List

logger = logging.getLogger("Phantom.AttackSystem")

class AutomatedAttackSystem:
    def __init__(self, ai_system, exploit_executor):
        self.ai = ai_system
        self.exploit_executor = exploit_executor
        self.nm = nmap.PortScanner()

    def _run_reconnaissance(self, campaign_id: str, target_ip: str) -> Dict[int, Dict]:
        """
        Performs deep service version detection using Nmap.
        Returns a map of {port: {product: str, version: str, cpe: str}}
        """
        logger.info(f"[*] Starting deep version detection on {target_ip}...")
        
        # -sV: Version detection, -T4: Aggressive timing, --script: SMB discovery
        args = "-sV -T4 --script=smb-os-discovery"
        self.nm.scan(target_ip, arguments=args)
        
        results = {}
        if target_ip in self.nm.all_hosts():
            for proto in self.nm[target_ip].all_protocols():
                ports = self.nm[target_ip][proto].keys()
                for port in ports:
                    svc = self.nm[target_ip][proto][port]
                    results[port] = {
                        "product": svc.get('product', ''),
                        "version": svc.get('version', ''),
                        "extrainfo": svc.get('extrainfo', ''),
                        "cpe": svc.get('cpe', '')
                    }
                    logger.info(f"  + Port {port}: {svc['product']} {svc['version']}")
        
        return results

    def _match_exploit_by_version(self, exploit_meta: Dict, detected_services: Dict) -> bool:
        """
        Validates if an exploit module is applicable based on version strings.
        """
        exploit_ports = exploit_meta.get('ports', [])
        for port in exploit_ports:
            if port in detected_services:
                svc = detected_services[port]
                # Priority: Check if version string matches known vulnerable patterns
                # This can be expanded with regex or CVE-lookup logic
                if svc['product'].lower() in exploit_meta['name'].lower():
                    return True
        return False

    async def process_target(self, target_ip, campaign_id):
        # 1. Version Detection Phase
        detected_services = self._run_reconnaissance(campaign_id, target_ip)
        
        # 2. Filter exploits based on version match
        # (This would integrate into your existing _run_campaign loop)
        # ...
